requirements.txt
openai
langchain-google-genai
pydantic_ai
pydantic
websockets
python-dotenv
